backend/requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
neo4j>=5.0.0
fastapi>=0.70.0
uvicorn>=0.15.0
pydantic>=2.0.0

# Data processing
numpy>=1.21.0
pandas>=1.3.0
wikipedia-api>=0.5.4
nltk>=3.8.1
tqdm>=4.65.0
requests>=2.31.0
wikiextractor==3.0.6

# Configuration
pyyaml>=6.0.0
python-dotenv>=0.19.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0 